pyproject.toml
[project]
name = "organize-photos"
version = "0.1.0"
description = "Organize your photos into meaningful structure based on timestamps"
authors = [{ name = "ohmycoffe", email = "ohmycoffe1@gmail.com" }]
dependencies = ["Pillow>=10.0.1", "click>=8.1.7"]
requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
[[tool.pdm.source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"

[tool.pdm.dev-dependencies]
dev = ["tox>=4.11.3", "pre-commit>=3.4.0"]
test = [
    "pytest",
    "numpy>=1.24.4",
    "piexif>=1.1.3",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.1",
    "pytest-mock>=3.11.1",
]
lint = ["flake8", "mypy", "flake8-pyproject>=1.2.3", "types-Pillow>=10.0.0.3"]
format = ["black", "isort"]


[tool.coverage.run]
omit = ["**/template_backport.py", "**/organize_photos/__main__.py"]

[tool.mypy]
[[tool.mypy.overrides]]
module = ["piexif"]
ignore_missing_imports = true

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"

[tool.bandit.assert_used]
skips = ['tests/**/*.py', 'tests/*.py']

[tool.codespell]
skip = 'htmlcov,tests/resources/valid_dirtree_recipe.json'
